requirements.txt
asgiref==3.2.5
astroid==2.4.2
autopep8==1.5
bleach==3.1.5
bottle==0.12.18
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coverage==5.1
defusedxml==0.6.0
demjson==2.2.4
Django==2.2
django-allauth==0.42.0
django-appconf==1.0.4
django-autoslug==1.9.7
django-ckeditor==5.9.0
django-compressor==2.4
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-haystack==3.0b2
django-ipware==2.1.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-password-reset==2.0
django-redis==4.11.0
django-taggit==1.3.0
django-uuslug==1.2.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
elasticsearch==7.7.1
elasticsearch-dsl==7.2.1
et-xmlfile==1.0.1
fastapi==0.60.1
idna==2.9
importlib-metadata==1.6.0
ipaddress==1.0.23
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.42.1
jsonpickle==1.4.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
markdown2==2.3.9
mccabe==0.6.1
mistune==0.8.4
mysqlclient==1.4.6
numpy==1.19.4
oauthlib==3.1.0
openpyxl==3.0.5
packaging==20.3
pandas==1.1.4
Pillow==7.1.2
pycodestyle==2.5.0
pydantic==1.6.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.2
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-logstash==0.4.6
python-memcached==1.59
python-slugify==4.0.0
# python3-openid==3.2.0
pytz==2020.1
raven==6.10.0
rcssmin==1.0.6
redis==3.5.2
requests==2.23.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.1
starlette==0.13.6
style==1.1.0
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
update==0.0.1
urllib3==1.25.9
webencodings==0.5.1
WeRoBot==1.12.0
Whoosh==2.7.4
wrapt==1.12.1
XlsxWriter==1.3.7
xmltodict==0.12.0
zipp==3.1.0

